Move list table inputs to element to be included in export
