--- packages/design-system/src/elements/toniq-list-table/list-table-inputs.ts
@@ -1,90 +1,0 @@
-import {ArrayElement, PartialAndUndefined, areJsonEqual} from '@augment-vir/common';
-import {CSSResult, HtmlInterpolation} from 'element-vir';
-
-export type ColumnsBase = ReadonlyArray<
-    Readonly<{
-        key: PropertyKey;
-        title: string;
-        disabled?: boolean;
-        option?: {
-            sticky?: boolean | undefined;
-            spaceEvenly?: boolean | undefined;
-        };
-        style?: CSSResult;
-    }>
->;
-
-export type ListTableRow<Columns extends ColumnsBase> = {
-    cells: Readonly<
-        ArrayElement<Columns> extends never
-            ? never
-            : Record<ArrayElement<Columns>['key'], HtmlInterpolation>
-    >;
-    rowActions?:
-        | PartialAndUndefined<{
-              click: (params: {
-                  clickEvent: MouseEvent;
-                  dispatch: (newEvent: Event) => void;
-              }) => void;
-          }>
-        | undefined;
-};
-
-export enum ListTableHeaderStyleEnum {
-    Default = 'default',
-    Transparent = 'transparent',
-}
-
-export type ListTableInputs = {
-    columns: Readonly<ColumnsBase>;
-    rows: ReadonlyArray<Readonly<ListTableRow<any>>>;
-    /** Defaults to ListTableHeaderStyleEnum.Default. */
-    headerStyle?: ListTableHeaderStyleEnum | undefined;
-    pagination?:
-        | Readonly<{
-              currentPage: number;
-              pageCount: number;
-          }>
-        | undefined;
-    showLoading?: boolean | undefined;
-    /**
-     * Used to show the table even if all items have not been painted yet, ideally to be used only
-     * in fixed/consistent column sizes in all rows
-     */
-    nonBlocking?: boolean | undefined;
-};
-
-export type CreateRowObjectCallback<EntryType, Columns extends ColumnsBase> = (
-    entry: Readonly<EntryType>,
-    index: number,
-) => Readonly<ListTableRow<Columns>>;
-
-export function createListTableTable<EntryType, const Columns extends ColumnsBase>({
-    entries,
-    columns,
-    createRowObject,
-}: {
-    entries: ReadonlyArray<Readonly<EntryType>>;
-    columns: Readonly<Columns>;
-    createRowObject: CreateRowObjectCallback<EntryType, Columns>;
-}): Pick<ListTableInputs, 'rows' | 'columns'> {
-    const columnsObject = Object.fromEntries(
-        columns.map((column) => [
-            column.key,
-            '',
-        ]),
-    );
-
-    const rows = entries.map((entry, index) => {
-        const row = createRowObject(entry, index);
-
-        if (!areJsonEqual(Object.keys(columnsObject).sort(), Object.keys(row.cells).sort())) {
-            console.error('broken list table row', {cells: row.cells, columns: columnsObject});
-            throw new Error('List table row keys does not match expect column keys.');
-        }
-
-        return row;
-    });
-
-    return {rows, columns};
-}
