+++ packages/design-system/src/elements/toniq-list-table/toniq-list-table.element.test.ts
@@ -1,10 +1,9 @@
 import {itCases} from '@augment-vir/browser-testing';
 import {fixture as renderFixture} from '@open-wc/testing';
 import {html} from 'element-vir';
-import {assertInstanceOf} from 'run-time-assertions';
+import {assertInstanceOf, assertTypeOf} from 'run-time-assertions';
 import {ToniqPagination} from '../toniq-pagination/toniq-pagination.element';
-import {createListTableTable} from './list-table-inputs';
-import {ToniqListTable} from './toniq-list-table.element';
+import {ToniqListTable, createListTableTable} from './toniq-list-table.element';
 
 const exampleListTableInputs = createListTableTable({
     entries: [
@@ -89,4 +88,60 @@ describe(ToniqListTable.tagName, () => {
             },
         },
     ]);
+});
+
+describe(createListTableTable.name, () => {
+    it('should have proper type enforcement', () => {
+        createListTableTable({
+            entries: [],
+            columns: [
+                {key: 'first', title: 'First'},
+                {key: 'second', title: 'Second'},
+            ],
+            // missing keys
+            // @ts-expect-error
+            createRowObject(entry) {
+                return {
+                    cells: {
+                        first: 'yo',
+                    },
+                };
+            },
+        });
+        createListTableTable({
+            entries: [
+                {
+                    key1: 'string',
+                    key2: 42,
+                },
+            ],
+            columns: [
+                {key: 'first', title: 'First'},
+                {key: 'second', title: 'Second'},
+            ],
+            createRowObject(entry) {
+                assertTypeOf(entry).toEqualTypeOf<Readonly<{key1: string; key2: number}>>();
+                return {
+                    cells: {
+                        first: 'yo',
+                        second: 'hi',
+                    },
+                };
+            },
+        });
+        createListTableTable({
+            entries: [
+                {
+                    key1: 'string',
+                    key2: 42,
+                },
+            ],
+            columns: [],
+            // can't do anything if columNames is empty
+            // @ts-expect-error
+            createRowObject(entry) {
+                return {cells: {}};
+            },
+        });
+    });
 });
